AG-13351 - Move <Collapsible /> component to shared website components
