+++ external/ag-website-shared/src/components/collapsible/Collapsible.tsx
