+++ external/ag-website-shared/src/components/site-header/HeaderNav.tsx
@@ -1,7 +1,7 @@
 import type { MenuItem } from '@ag-grid-types';
+import { Collapsible } from '@ag-website-shared/components/collapsible/Collapsible';
 import { Icon } from '@ag-website-shared/components/icon/Icon';
 import MenuIcon from '@ag-website-shared/images/inline-svgs/menu-icon.svg?react';
-import { Collapsible } from '@components/Collapsible';
 import { getPageNameFromPath } from '@components/docs/utils/urlPaths';
 import { useFrameworkFromStore } from '@utils/hooks/useFrameworkFromStore';
 import { urlWithPrefix } from '@utils/urlWithPrefix';
